The ring buffer code has been placed in the simtemp_ringbuf module
